hackyeahAPI/requirements.txt
django
djangorestframework
psycopg2-binary
daphne
pandas
numpy
whitenoise
django-environ
gunicorn
djangorestframework-simplejwt
